refactor /-/v1/instances route
